LPS-59536 Move to seperate package
