+++ modules/portal-cache/portal-cache-ehcache/src/main/java/com/liferay/portal/cache/ehcache/internal/MultiVMEhcachePortalCacheManager.java
@@ -17,6 +17,7 @@ package com.liferay.portal.cache.ehcache.internal;
 import com.liferay.portal.cache.PortalCacheBootstrapLoaderFactory;
 import com.liferay.portal.cache.PortalCacheListenerFactory;
 import com.liferay.portal.cache.PortalCacheManagerListenerFactory;
+import com.liferay.portal.cache.ehcache.internal.configurator.AbstractEhcachePortalCacheManagerConfigurator;
 import com.liferay.portal.kernel.cache.PortalCacheManager;
 import com.liferay.portal.kernel.cache.PortalCacheManagerNames;
 import com.liferay.portal.kernel.cache.configurator.PortalCacheConfiguratorSettings;
