+++ modules/portal-cache/portal-cache-ehcache/src/main/java/com/liferay/portal/cache/ehcache/internal/configurator/MultiVMEhcachePortalCacheManagerConfigurator.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache.ehcache.internal;
+package com.liferay.portal.cache.ehcache.internal.configurator;
 
 import com.liferay.portal.cache.PortalCacheReplicator;
 import com.liferay.portal.cache.configuration.PortalCacheConfiguration;
