+++ modules/portal-cache/portal-cache-ehcache/src/main/java/com/liferay/portal/cache/ehcache/internal/configurator/SingleVMEhcachePortalCacheManagerConfigurator.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache.ehcache.internal;
+package com.liferay.portal.cache.ehcache.internal.configurator;
 
 import com.liferay.portal.kernel.cache.PortalCacheManager;
 import com.liferay.portal.kernel.cache.PortalCacheManagerNames;
